+++ src/bpp/admin/uczelnia.py
@@ -89,6 +89,7 @@ class UczelniaAdmin(
                     "pbn_api_user",
                     "pbn_api_kasuj_przed_wysylka",
                     "pbn_api_nie_wysylaj_prac_bez_pk",
+                    "pbn_api_afiliacja_zawsze_na_uczelnie",
                 ),
             },
         ),